Simplify testing method, and WhereClause.
